third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/cross-origin-opener-policy/reporting/navigation-reporting/reporting-coop-navigated-opener.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigation-report-from-opener-navigation
  assert_equals: expected (string) "coop" but got (undefined) undefined
Harness: the test ran to completion.
